+++ include/calculation.h
@@ -7,7 +7,7 @@
 #include "pico/fft.h"
 #define SAMPLING_INTERVAL_US (100)
 #pragma once
-float calculate_frequency(uint8_t *capture_buf);
+float calculate_frequency(uint16_t *capture_buf);
 void precalculate_window();
 
 #endif
